Update graphql schemas, separate between university and country schema
